Fixed credential validation to actually work; still need to implement a legitimate one-way hash

+++ src/utils/User.java
@@ -1,15 +1,15 @@
 package utils;
 
 public class User {
-	private int userId;
+	private long userId;
 	private String username;
 	
-	public User(int id, String name) {
+	public User(long id, String name) {
 		this.userId = id;
 		this.username = name;
 	}
 	
-	public int getUserId() {
+	public long getUserId() {
 		return userId;
 	}
